[hebe-jvm] Move generators to separate file
